cost-tools/requirements.txt
azure-identity
azure-mgmt-compute
azure-mgmt-subscription
tabulate
